+++ rosflight/src/mavrosflight/time_manager.cpp
@@ -80,7 +80,7 @@ void TimeManager::handle_mavlink_message(const mavlink_message_t &msg)
   }
 }
 
-ros::Time TimeManager::get_ros_time_ms(uint64_t boot_ms)
+ros::Time TimeManager::get_ros_time_ms(uint32_t boot_ms)
 {
   if (!initialized_)
     return ros::Time::now();
